chore: Update dependencies
Removed the dependancy list from package.json so that foundation
is the source truth. Keeping package.json to have the name for
the project and the scripts.

Update the dependancy base image for the project and only chown
files that actually need to be chown (i.e., not node_modules
folder).